Update and rename TransferTypes.py to GenerateFile.py
